requirements-dev.txt
django-debug-toolbar==2.0
coverage==4.5.4
pycodestyle==2.5.0
locustio==0.11.0
